decoupl daily email digest subscripiton functionality 🐿 v2.12.5
