+++ src/components/bottom/oneclick-daily-digest/main.js
@@ -1,5 +1,7 @@
-import myft from 'next-myft-client';
-import Cookies from 'js-cookie';
+// import myft from 'next-myft-client';
+// import Cookies from 'js-cookie';
+
+import signUpToDigest from './signUp-to-digest';
 
 module.exports = function customSetup (banner, done) {
 	if (window.FT && window.FT.flags && window.FT.flags.oneClickDailyDigest) {
@@ -9,37 +11,20 @@ module.exports = function customSetup (banner, done) {
 			const closeButton = banner.bannerElement.querySelector('.o-banner__close');
 			closeButton.classList.add('o-banner__close-color');
 			const signupContent = banner.bannerElement.querySelector('.o-banner_signup-content');
-			signupContent.classList.add('is-visible');
+			signupContent.classList.add('o-banner_is-visible');
 			const successContent = banner.bannerElement.querySelector('.o-banner_success-content');
-			successContent.classList.remove('is-visible');
+			successContent.classList.remove('o-banner_is-visible');
 		}
 
 		function handleSignUpClick (evt) {
 			evt.preventDefault();
-
-			function getCSRFToken () {
-				const desiredTokenLength = 36;
-				const token = Cookies.get('FTSession_s') || Cookies.get('FTSession');
-				const csrfToken = token ? token.slice(-desiredTokenLength) : '';
-				return csrfToken;
-			}
-
-			function addUserToDigest () {
-				return myft.add('user', null, 'preferred', 'preference', 'email-digest', {
-					token: getCSRFToken,
-					_rel:{
-						type: 'daily',
-						sendTime:'every morning'
-					}
-				});
-			}
 
 			function logError (err) {
 				/* eslint no-console:0 */
 				console.log({ info: 'could not add user to the daily digest', err });
 			}
 
-			myft.init().then(addUserToDigest).then(loadSuccessContent).catch(logError);
+			signUpToDigest().then(loadSuccessContent).catch(logError);
 			return false;
 		}
