+++ src/components/bottom/oneclick-daily-digest/main.scss
@@ -41,7 +41,7 @@
 	font-size: 16px;
 }
 
-.is-visible {
+.o-banner_is-visible {
 	display: none;
 }
 .o-banner_button-signup {
