+++ templates/partials/bottom/oneclick-daily-digest.html
@@ -14,7 +14,7 @@
 				</div>
 			</div>
 		</div>
-		<div class="o-banner_success-content is-visible">
+		<div class="o-banner_success-content o-banner_is-visible">
 			<div class="o-banner__content">
 				<span class="o-banner_success-icon-tick"></span>
 				<strong>Great! You're all set for the daily Email Digest.</strong>
